chore: move bash tool execution to browser tool
